Feat: 1주차 코드 리뷰 반영 (Snoop)
