+++ fe/src/components/commons/Icon/Icon.tsx
@@ -1,16 +1,12 @@
 import * as S from './IconStyle';
 import { icons } from '@assets/icons';
-
-export interface IconStyleProps {
-  size?: number;
-  color?: string;
-}
+import { IconStyleProps } from './IconStyle';
 
 interface IconProps extends IconStyleProps {
   name: keyof typeof icons;
 }
 
-export const Icon = ({ name = 'arrowDown', size = 16, color }: IconProps) => {
+export const Icon = ({ name, size = 16, color = 'neutralText' }: IconProps) => {
   return (
     <S.Icon size={size} color={color}>
       {icons[name]}
