+++ fe/src/components/commons/Tab/Tab.tsx
@@ -1,6 +1,4 @@
-import { ReactElement } from 'react';
 import * as S from './TabStyle';
-import { Icon } from '@components/commons/index';
 
 interface OptionType {
   name: string;
@@ -8,16 +6,15 @@ interface OptionType {
 }
 
 interface TabProps {
-  children: ReactElement | string;
   onClick: () => void;
   activeId: string;
   options: OptionType[];
 }
 
-export const Tab = ({ children, onClick, activeId, options }: TabProps) => {
+export const Tab = ({ onClick, activeId, options }: TabProps) => {
   return (
     <S.Wrap>
-      {options?.map((option) => (
+      {options.map((option) => (
         <S.Tab
           key={option.id}
           onClick={onClick}
