+++ fe/src/styles/index.css
@@ -1,4 +1,4 @@
 @font-face {
   font-family: 'SF Pro';
-  src: url('src/assets/fonts/SF-Pro.ttf') format('truetype');
+  src: url('src/assets/fonts/SF-Pro.woff2') format('woff2');
 }
